Make ClientObj, SlaveObj, and ConnectionHandler inner classes of Server
